Reworked the hash function for the position.

+++ Assets/Scripts/Runtime/MarchingCubes/Tools/ChunksMath.cs
@@ -11,7 +11,7 @@ namespace MarchingCubesProject.Tools
     public static class ChunksMath
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static NativeHashMap<long, IntPtr> GetChunksDataPointersInsideArea(
+        public static NativeHashMap<int, IntPtr> GetChunksDataPointersInsideArea(
             Area area, Vector3Int chunksSize, IChunksContainer chunksContainer)
         {
             Vector3Int affectedAreaSize = area.Parallelepiped.Size;
@@ -20,8 +20,8 @@ namespace MarchingCubesProject.Tools
             int maxZChunks = affectedAreaSize.z / chunksSize.z + 1;
             int maxAffectedChunksCount = maxXChunks * maxYChunks * maxZChunks;
 
-            NativeHashMap<long, IntPtr> pointers
-                = new NativeHashMap<long, IntPtr>(maxAffectedChunksCount, Allocator.Persistent);
+            NativeHashMap<int, IntPtr> pointers
+                = new NativeHashMap<int, IntPtr>(maxAffectedChunksCount, Allocator.Persistent);
 
             for (int y = Mathf.FloorToInt((float)area.Min.y / chunksSize.y) * chunksSize.y; y < area.Max.y; y += chunksSize.y)
             {
@@ -32,7 +32,7 @@ namespace MarchingCubesProject.Tools
                         int localChunkX = x / chunksSize.x;
                         int localChunkY = y / chunksSize.y;
                         int localChunkZ = z / chunksSize.z;
-                        long positionHash = PositionHasher.GetPositionHash(localChunkX, localChunkY, localChunkZ);
+                        int positionHash = PositionHasher.GetPositionHash(localChunkX, localChunkY, localChunkZ);
                         IChunk chunk = chunksContainer.GetChunk(positionHash);
 
                         if (chunk == null)
